Added Thinkpad T530 Configuration and Scripts
The configuration files and scripts added provide support
for docking and undocking of my (and your) Lenovo
Thinkpad T530 with the proprietary nVidia drivers. These
will likely need to be catered to fit your system, but
they'll give you the general idea.

Also included is a small list of yum repositories I use.